# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Matka Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/signup:
    post:
      operationId: userSignup
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Phone already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getAdminMe
      tags: [auth]
      summary: Get current admin info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        admin:
          $ref: "#/components/schemas/Admin"
      required:
        - token
        - admin
    SignupRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        password:
          type: string
      required:
        - name
        - phone
        - password
    SignupResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    Admin:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - createdAt
    DashboardStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalBidsToday:
          type: integer
        totalProfit:
          type: number
        activeMarkets:
          type: integer
        depositsToday:
          type: number
        withdrawalsToday:
          type: number
        winAmount:
          type: number
        lossAmount:
          type: number
        netWinLoss:
          type: number
        recentBids:
          type: array
          items:
            $ref: "#/components/schemas/Bid"
      required:
        - totalUsers
        - totalBidsToday
        - totalProfit
        - activeMarkets
        - depositsToday
        - withdrawalsToday
        - winAmount
        - lossAmount
        - netWinLoss
        - recentBids
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        walletBalance:
          type: number
        isBlocked:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - walletBalance
        - isBlocked
        - createdAt
...
    Bid:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        marketId:
          type: integer
        marketName:
          type: string
        gameType:
          type: string
        amount:
          type: number
        digit:
          type: string
        openTime:
          type: string
        closeTime:
          type: string
        currentTime:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - marketId
        - marketName
        - gameType
        - amount
        - digit
        - status
        - createdAt
...