# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * Matka Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { Bid } from './bid';

export interface DashboardStats {
  totalUsers: number;
  totalBidsToday: number;
  totalProfit: number;
  activeMarkets: number;
  depositsToday: number;
  withdrawalsToday: number;
  winAmount: number;
  lossAmount: number;
  netWinLoss: number;
  recentBids: Bid[];
}